web-animations/animation-group/animation-group-clone.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationGroup clone() method test</title>
<meta name="assert" content="Creates a deep copy of this AnimationGroup object">
<link rel="help" href="w3c.github.io/web-animations/#the-start-time-of-children-of-an-animation-group">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var source = new AnimationGroup([]);
    var cloned = source.clone();

    assert_false(cloned === source,
        'Cloned AnimationGroup object should be a copy of source object');
    assert_object_equals(cloned, source, 'Cloned AnimationGroup object should be ' +
        'identical to source object');

    assert_false(cloned.children === source.children, 'Cloned AnimationGroup.children ' +
        'should not be the same object as source.children');
    assert_array_equals(cloned.children, [], 'Cloned AnimationGroup.children should be empty');

    assert_false(cloned.timing === source.timing, 'Cloned AnimationGroup.timing ' +
        'should not be the same object as source.timing');
    assert_object_equals(cloned.timing, source.timing, 'Cloned AnimationGroup.timing ' +
        'should be identical to source.timing');
}, 'Test AnimationGroup.clone() creates a copy of AnimationGroup object without children ' +
    'and with default timing');

test(function() {
    var source = new AnimationGroup([], 5);
    var cloned = source.clone();

    assert_false(cloned === source,
        'Cloned AnimationGroup object should be a copy of source object');
    assert_object_equals(cloned, source, 'Cloned AnimationGroup object should be ' +
        'identical to source object');

    assert_false(cloned.children === source.children, 'Cloned AnimationGroup.children ' +
        'should not be the same object as source.children');
    assert_array_equals(cloned.children, [], 'Cloned AnimationGroup.children should be empty');

    assert_false(cloned.timing === source.timing, 'Cloned AnimationGroup.timing ' +
        'should not be the same object as source.timing');
    assert_object_equals(cloned.timing, source.timing, 'Cloned AnimationGroup.timing ' +
        'should be identical to source.timing');
}, 'Test AnimationGroup.clone() creates a copy of AnimationGroup object without children ' +
    'and with non-default timing duration');

test(function() {
    var timing = {
        delay: 1,
        endDelay: 1,
        fill: 'none',
        iterationStart: 1.1,
        iterations: 2.0,
        duration: 3,
        playbackRate: -2,
        playbackDirection: 'reverse',
        direction: 'reverse',
        easing: 'ease'};
    var source = new AnimationGroup([], timing);
    var cloned = source.clone();

    assert_false(cloned === source,
        'Cloned AnimationGroup object should be a copy of source object');
    assert_object_equals(cloned, source, 'Cloned AnimationGroup object should be ' +
        'identical to source object');

    assert_false(cloned.children === source.children, 'Cloned AnimationGroup.children ' +
        'should not be the same object as source.children');
    assert_array_equals(cloned.children, [], 'Cloned AnimationGroup.children should be empty');

    assert_false(cloned.timing === source.timing, 'Cloned AnimationGroup.timing ' +
        'should not be the same object as source.timing');
    assert_object_equals(cloned.timing, source.timing, 'Cloned AnimationGroup.timing ' +
        'should be identical to source.timing');
}, 'Test AnimationGroup.clone() creates a copy of AnimationGroup object without children ' +
    'and with non-default timing');

test(function() {
    var node = newAnimation(createDiv(this));
    var children = [node];
    var source = new AnimationGroup(children);
    var cloned = source.clone();

    assert_false(cloned === source,
        'Cloned AnimationGroup object should be a copy of source object');
    assert_object_equals(cloned, source, 'Cloned AnimationGroup object should be ' +
        'identical to source object');

    assert_false(cloned.children === source.children, 'Cloned AnimationGroup.children ' +
        'should not be the same object as source.children');
    assert_equals(cloned.children.length, source.children.length, 'Cloned AnimationGroup ' +
        'object should contain the same number of children');
    assert_false(cloned.clidren[0] === source.children[0], 'First child of cloned ' +
        'AnimationGroup object should not be the same as first child of source');
    assert_object_equals(cloned.clidren[0], source.children[0], 'First child of cloned ' +
        'AnimationGroup object should be identical to first child of source');
}, 'Test AnimationGroup.clone() creates a copy of AnimationGroup object including children');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = new AnimationGroup([node1,node2]);
    var node4 = newAnimation(createDiv(this));
    var node5 = newAnimation(createDiv(this));
    var node6 = new AnimationSequence([node4, node5]);

    var source = new AnimationGroup([node3, node6]);
    var cloned = source.clone();

    assert_false(cloned === source,
        'Cloned AnimationGroup object should be a copy of source object');
    assert_object_equals(cloned, source, 'Cloned AnimationGroup object should be ' +
        'identical to source object');

    var expected = [node1, node2, node3, node4, node5, node6];
    var actual = [
        cloned.children[0].children[0],
        cloned.children[0].children[1],
        cloned.children[0],
        cloned.children[1].children[0],
        cloned.children[1].children[1],
        cloned.children[1]
    ];
    for (var i = 0; i < expected.length; i++) {
        assert_false(actual[i] === expected[i], 'Descendant of cloned AnimationGroup ' +
            'object should be a copy of corresponding descendant of source object');
        assert_object_equals(actual[i], expected[i], 'Descendant of cloned AnimationGroup ' +
            'object should be identical to corresponding descendant of source object');
    }
    assert_false(cloned.timing === source.timing, 'Cloned AnimationGroup.timing ' +
        'should not be the same object as source.timing');
    assert_object_equals(cloned.timing, source.timing, 'Cloned AnimationGroup.timing ' +
        'should be identical to source.timing');
}, 'Test AnimationGroup.clone() creates a deep copy of AnimationGroup object');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = new AnimationGroup([node1]);

    var cloned = node1.clone();
    assert_equals(cloned.parent, null, 'Cloned AnimationGroup.parent should be null');
    assert_equals(node1.parent, node2, 'Source AnimationGroup.parent should not be affected');
}, 'Test AnimationGroup.clone() set parent attribute to null on cloned object');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = new AnimationGroup([]);
    var node3 = new AnimationGroup([node1, node2]);

    var cloned = node2.clone();
    assert_equals(cloned.previousSibling, null,
        'Cloned AnimationGroup.previousSibling should be null');
    assert_equals(node2.previousSibling, node1,
        'Source AnimationGroup.previousSibling should not be affected');
}, 'Test AnimationGroup.clone() set previousSibling attribute to null on cloned object');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = newAnimation(createDiv(this));
    var node3 = new AnimationGroup([node1, node2]);

    var cloned = node1.clone();
    assert_equals(cloned.nextSibling, null,
        'Cloned AnimationGroup.nextSibling should be null');
    assert_equals(node1.nextSibling, node2,
        'Source AnimationGroup.nextSibling should not be affected');
}, 'Test AnimationGroup.clone() set nextSibling attribute to null on cloned object');
</script>
</body>
